Update items.sql
added an additional filed called "updated_at" to be used as a ETag to check the version of the data by tracking the last modification time of each item. 

It is set as the current timestamp automatically when the row is first inserted and will be automatically updates it whenever any column in the row changes (INSERT, UPDATE).